Generalized rule class for NSG rule
