+++ installer/resources/lambda_rule_engine/files/rule_engine_cloudwatch_rules.json
@@ -3613,15 +3613,15 @@
     "ruleUUID": "azure_enable_network_security_SQLSERVER",
     "policyId": "PacMan_Azure_Enable_Network_Security_for_SQLSERVER",
     "ruleName": "Enable_Network_Security_for_SQLSERVER",
-    "targetType": "sqlserver",
+    "targetType": "nsg",
    "assetGroup": "Azure",
     "alexaKeyword": "Enable Network Security for SQLSERVER",
-    "ruleParams": "{\"params\":[{\"encrypt\":false,\"value\":\"check-for-azure-security-rule-for-postgresql\",\"key\":\"ruleKey\"},{\"key\":\"protocol\",\"value\":\"tcp\",\"isValueNew\":true,\"encrypt\":false},{\"key\":\"port\",\"value\":\"1433\",\"isValueNew\":true,\"encrypt\":false},{\"key\":\"policyName\",\"value\":\"Access@should@be@restricted@for@permissive@Network@Security@Groups@with@Internet-facing@SQLSERVER\",\"isValueNew\":true,\"encrypt\":false},{\"encrypt\":false,\"value\":\"high\",\"key\":\"severity\"},{\"encrypt\":false,\"value\":\"security\",\"key\":\"ruleCategory\"},{\"encrypt\":false,\"value\":\"\",\"key\":\"ruleOwner\"}],\"environmentVariables\":[],\"ruleId\":\"PacMan_Azure_Enable_Network_Security_to_SQLSERVER\",\"autofix\":false,\"alexaKeyword\":\"Enable Network Security for SQLSERVER\",\"ruleRestUrl\":\"\",\"targetType\":\"sqlserver\",\"pac_ds\":\"azure\",\"policyId\":\"PacMan_Azure_Enable_Network_Security_to_SQLSERVER\",\"assetGroup\":\"azure\",\"ruleUUID\":\"azure_enable_network_security_SQLSERVER\",\"ruleType\":\"ManageRule\"}",
+    "ruleParams": "{\"params\":[{\"encrypt\":false,\"value\":\"check-for-azure-nsg-rule\",\"key\":\"ruleKey\"},{\"key\":\"protocol\",\"value\":\"tcp\",\"isValueNew\":true,\"encrypt\":false},{\"key\":\"port\",\"value\":\"1433\",\"isValueNew\":true,\"encrypt\":false},{\"key\":\"policyName\",\"value\":\"Restrict TCP for 1433 to trust IPs only on SQL servers\",\"isValueNew\":true,\"encrypt\":false},{\"encrypt\":false,\"value\":\"high\",\"key\":\"severity\"},{\"encrypt\":false,\"value\":\"security\",\"key\":\"ruleCategory\"},{\"encrypt\":false,\"value\":\"\",\"key\":\"ruleOwner\"}],\"environmentVariables\":[],\"ruleId\":\"PacMan_Azure_Enable_Network_Security_for_SQLSERVER\",\"autofix\":false,\"alexaKeyword\":\"Enable Network Security for SQLSERVER\",\"ruleRestUrl\":\"\",\"targetType\":\"sqlserver\",\"pac_ds\":\"azure\",\"policyId\":\"PacMan_Azure_Enable_Network_Security_for_SQLSERVER\",\"assetGroup\":\"azure\",\"ruleUUID\":\"azure_enable_network_security_SQLSERVER\",\"ruleType\":\"ManageRule\"}",
     "ruleFrequency": "0 0/6 * * ? *",
     "ruleExecutable": "",
     "ruleRestUrl": "",
     "ruleType": "ManageRule",
-    "ruleArn": "arn:aws:events:us-east-1:***REMOVED***:rule/aazure_enable_network_security_SQLSERVER",
+    "ruleArn": "arn:aws:events:us-east-1:***REMOVED***:rule/azure_enable_network_security_SQLSERVER",
     "status": "ENABLED",
     "userId": "ASGC",
     "displayName": "Enable Network Security Groups on sql server",
@@ -3630,6 +3630,7 @@
     "severity": "high",
     "category": "security"
   },
+
    {
     "ruleId": "PacMan_Unrestricted_Microsoft_SQL_Database_Access",
     "ruleUUID": "unrestricted_Microsoft_SQL_Database_Access",
